API documentation

Kanones.code — Function

Compose a digital code for adj.

code(verb)

Compose a digital code for inf.

code(inf)

Compose digital code for ptcpl.

code(ptcpl)

Compose a digital code for vadj.

code(vadj)

Compose a digital code for noun.

code(noun)

Compose a digital code string for the form identified in rule.

code(rule)

Compose a digital code for adj.

code(adj)

Compose digital code for morphological form identified in rule.

code(rule)

Compose a digital code for adv.

code(adv)

Compose a digital code for pronoun.

code(pronoun)

Compose digital code for uform.

code(uform)

Kanones.formurn — Function

Compose a FormUrn for a GMFFiniteVerb.

formurn(verbform)

Compose a FormUrn for an GMFInfinitive.

formurn(infinitive)

Compose a FormUrn for an GMFParticiple.

formurn(ptcpl)

Compose a FormUrn for an GMFVerbalAdjective.

formurn(vadj)

Compose a FormUrn for a GMFNoun.

formurn(noun)

Compose an abbreviated URN for a rule from a NounRule.

formurn(rule)

Compose a FormUrn for an GMFAdjective.

formurn(adj)

Compose an abbreviated URN for a morphological form identified in rule.

formurn(rule)

Compose a FormUrn for an GMFAdjective.

formurn(adv)

Compose a FormUrn for a GMFPronoun.

formurn(pronoun)

Compose a FormUrn for an GMFUninflected.

formurn(uninflected)

Parsing

Kanones.generate — Function

Generate a form for a given stem and rule.

generate(stem, rule; ortho)

Generate a form for a given noun stem and rule by combining stem and ending, then adding appropriate accent for this lexical item in this form, and finally stripping off metadata characters marking vowel quantity and morpheme boundaries.

generate(stem, rule; ortho)

Kanones.decline — Function

Decline all case-number combinations of lex, a noun.

decline(lex, kd; withvocative)