Julia language API documentation

Julia language API documentation

Morphology

Structures

Functions

Datasets

Structures

Functions

Analysis

Exported functions

Functions used internally in parsing via a FST

Generation

Kanones.FstBuilder

Morphology

Structures

Kanones.MorphologicalForm — Type

Abstract type of a morphological form in Kanones.

Kanones.UninflectedForm — Type

Uninflected forms have a single property: the "part of speech".

Kanones.NounForm — Type

Nouns have gender, case and number.

Kanones.FiniteVerbForm — Type

Finite verbs have person, number, tense, mood and voice.

Functions

CitableBase.urn — Function

Delegate to specific functions based on type's citable trait value.

urn(x)

It is an error to invoke the urn function on material that is not citable.

urn(_, x)

Citable text content should implement cex.

urn(_, txt)

Citable content should implement cex.

urn(_, obj)

URN identifyiing psg.

urn(psg)

Required function for Citable abstraction.

URN identifyiing doc.

urn(doc)

Required function for Citable abstraction.

Implement urn function required by Citable interface for CatalogedText.

urn(cataloged)

Unique identifier for AnalyzedToken (required for Citable interface).

urn(at)

Convert a MorphologicalForm to a Cite2Urn.

urn(mf)

All subclasses of MorphologicalForm should implement this specifically for their subclass.

Compose URN for an UninflectedForm.

urn(uform)

Compose a Cite2Urn for a NounForm.

urn(noun)

Compose a Cite2Urn for an AdjectiveForm.

urn(adj)

Compose URN for a FiniteVerbForm.

urn(verb)

CitableBase.cex — Function

Delegate to specific functions based on type's citable trait value.

cex(x)
cex(x, delim)

It is an error to invoke the cex function on material that is not citable.

cex(_, x, delim)

Citable text content should implement cex.

cex(_, txt, delim)

Citable content should implement cex.

cex(_, obj, delim)

Format a CitablePassage as a delimited-text string.

cex(psg)
cex(psg, delim)

Required function for Citable abstraction.

Format a CitableDocument as a CEX ctsdata block.

cex(doc)
cex(doc, delim)

Required function for Citable abstraction.

Compose a delimited-text string for a corpus.

cex(c; delimiter)

Implement cex function required by Citable interface for CatalogedText.

cex(cataloged)
cex(cataloged, delimiter)

Serialize an AnalyzedToken as delimited text (required for Citable interface).

cex(at)
cex(at, delim)

Uses abbreviated URNs. These can be expanded to full CITE2 URNs when read back with a URN registry, or the delimited function can be used with a URN registry to write full CITE2 URNs.

Convert a MorphologicalForm form to a delimited-text string.

cex(mf, delim)

All subclasses of MorphologicalForm should implement this specifically for their subclass.

Compose URN for an UninflectedForm.

cex(uform)

Compose CEX representation for a NounForm.

cex(noun)
cex(noun, delim)

Compose CEX representation for a AdjectiveForm.

cex(adj)
cex(adj, delim)

Compose CEX representation for a FiniteVerbForm.

cex(verb)

Kanones.morphform — Function

Create a MorphologicalForm from a Cite2Urn value.

morphform(urn)

Create a MorphologicalForm from a FormUrn.

morphform(u)

Create a MorphologicalForm from an Analysis.

morphform(a)

Create a MorphologicalForm from an string value.

morphform(code)

Kanones.abbrurn — Function

Generic function to convert form information in a Rule to a RuleUrn.

abbrurn(rule)

Compose an abbreviated URN for a form from a NounRule.

abbrurn(rule)

Compose an abbreviated URN for a form from a AdjectiveRule.

abbrurn(rule)

Kanones.formscex — Function

Compose CEX representation of CITE colletion of all recognized forms.

formscex()

Currently, includes only:

unanalyzed forms

uninflected forms

noun forms

inflected verb forms

Kanones.formurn — Function

Generic function to convert form information in a MorphologicalForm to a FormUrn.

formurn(rule)

Compose a FormUrn for an UninflectedForm.

formurn(uninflected)

Compose a FormUrn for a NounForm.

formurn(nounform)

Compose a FormUrn for a PronounForm.

formurn(pronounform)

Compose a FormUrn for an AdjectiveForm.

formurn(adj)

Compose a FormUrn for a FiniteVerbForm.

formurn(verbform)

Compose a FormUrn for an InfinitiveForm.

formurn(infinitive)

Compose a FormUrn for an InfinitiveForm.

formurn(ptcpl)

Compose a FormUrn for an InfinitiveForm.

formurn(vadj)

Kanones.pospairs — Constant

Constants for analytical types ("parts of speech").

Kanones.uninflectedpairs — Constant

Constants for uninflected type of uninflected form.

Kanones.casepairs — Constant

Constants for substantive case.

Kanones.personpairs — Constant

Constants for persons.

Kanones.numberpairs — Constant

Constants for number.

Kanones.voicepairs — Constant

Constants for voice.

Kanones.tensepairs — Constant

Constants for tense.

Kanones.moodpairs — Constant

Constants for mood.

Kanones.valuedict — Function

Create a dictionary keyed by the value of a label-value pair.

valuedict(prs)

Kanones.labeldict — Function

Create a dictionary keyed by the label of a label-value pair.

labeldict(prs)

Kanones.labelform — Function

Compose a human-readable lable for a morphological form identifier.

labelform(s)

Compose a human-readable label for a FormUrn.

labelform(furn)

Compose a human-readable label for a FormUrn.

labelform(furn)

Kanones.nounform — Function

Create a NounForm from a string value.

nounform(code)

Create a NounForm from a Cite2Urn.

nounform(urn)

Create a NounForm from a FormUrn.

nounform(f)

Create a NounForm from an Analysis.

nounform(a)

Kanones.nounscex — Function

Compose CEX representation of URNs and labels for noun forms.

nounscex()

Kanones.uninflectedform — Function

Create UninflectedForm from a Cite2Urn.

uninflectedform(urn)

Create UninflectedForm from a FormUrn.

uninflectedform(u)

Create UninflectedForm from a string value.

uninflectedform(codeString)

Create UninflectedForm from a Char.

uninflectedform(ch)

Create UninflectedForm from an Int.

uninflectedform(code)

Create UninflectedForm from an Analysis.

uninflectedform(a)

Kanones.uninflectedcex — Function

Compose CEX representation of URNs and labels for uninflected forms.

uninflectedcex()

Kanones.finiteverbform — Function

Create a FiniteVerbForm from a string value.

finiteverbform(code)

Create a FiniteVerbForm from a Cite2URN.

finiteverbform(urn)

Create a FiniteVerbForm from a FormUrn.

finiteverbform(f)

Create a FiniteVerbForm from an Analysis.

finiteverbform(a)

Datasets

Structures

Kanones.Dataset — Type

A Kanones dataset in a local file system.

Members

dirs is a list of full paths to the root directory of one or more datasets.

Their subdirectory organization must follow Kanones' specifications.

orthography is an implementation of the GreekOrthography interface

Kanones.KanonesIO — Type

Implementations of this type can read delimited and write fst.

Kanones.NounParser — Type

KanonesIO type for reading and writing data for noun forms.

Kanones.NounStem — Type

A record for a single noun stem.

Kanones.NounRule — Type

Inflectional rule for uninflected lexical items.

Kanones.UninflectedParser — Type

KanonesIO type for reading and writing stem for uninflected forms.

Kanones.UninflectedStem — Type

A record for a single uninflected stem.

Kanones.UninflectedRule — Type

Inflectional rule for uninflected lexical items.

Kanones.VerbParser — Type

KanonesIO type for reading and writing data for finite verb form.

Functions

Exported:

Kanones.dataset — Function

Create a Kanones.Dataset from a single data source.

dataset(dir; ortho)

Arguments

dir Full path to a directory with Kanones data.

ortho An instance of a GreekOrthography; defaults to LiteraryGreekOrthography.

Create a Kanones.Dataset from one or more data sources.

dataset(srclist; ortho)

Arguments

srclist List of full paths to a directory with Kanones data.

ortho An instance of a GreekOrthography; defaults to LiteraryGreekOrthography.

Kanones.rulesarray — Function

Read all rules data from a list of directories into an array of Rules.

rulesarray(kd)

Read all rules data from a Kanones.Dataset into an array of Rules.

rulesarray(dirlist)

Kanones.stemsarray — Function

Read all stem data from a list of directories into an array of Stems.

stemsarray(kd)

Read all stem data from a Kanones.Dataset into an array of Stems.

stemsarray(dirlist)

Internals:

Kanones.readstemrow — Function

Implementation of reading one row of a stems table for uninflected tokens.

readstemrow(usp, delimited)
readstemrow(usp, delimited, delimiter)

Read one row of a stems table for noun tokens and create a NounStem.

readstemrow(usp, delimited)
readstemrow(usp, delimited, delimiter)

Read one row of a stems table for noun tokens and create a NounStem.

readstemrow(usp, delimited)
readstemrow(usp, delimited, delimiter)

Read one row of a stems table for noun tokens and create a NounStem.

readstemrow(usp, delimited)
readstemrow(usp, delimited, delimiter)

Implementation of reading one row of a stems table for finite verbs.

readstemrow(usp, delimited)
readstemrow(usp, delimited, delimiter)

Read one row of a stems table for adjective tokens and create an AdjectiveStem.

readstemrow(usp, delimited)
readstemrow(usp, delimited, delimiter)

Kanones.readrulerow — Function

Implementation of reading one row of a rules table for uninflected tokens.

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Implementation of reading one row of a rules table for uninflected tokens.

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Read one row of a stems table for noun tokens and create a NounStem.

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Implementation of reading one row of a rules table for irregular tokens.

readrulerow(ruleparser, delimited)
readrulerow(ruleparser, delimited, delimiter)

Read one row of a rules table for verb tokens and create a FiniteVerbRule

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Read one row of a rules table for infinitives and create an InfinitiveRule.

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Read one row of a rules table for infinitives and create an InfinitiveRule.

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Read one row of a rules table for infinitives and create an InfinitiveRule.

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Implementation of reading one row of a rules table for uninflected tokens.

readrulerow(usp, delimited)
readrulerow(usp, delimited, delimiter)

Kanones.id — Function

Identify identifier URN for an UninflectedStem.

id(uninfl)

Identify identifier URN for an UninflectedRule.

id(uninfl)

Identify identifier URN for a NounStem.

id(n)

Identify identifier URN for a NounRule.

id(n)

Kanones.lexeme — Function

Identify lexeme URN for an UninflectedStem.

lexeme(uninfl)

Identify lexeme URN for a NounStem.

lexeme(n)

Kanones.abbrformurn — Function

Compose FormUrn for an irregular noun stem.

abbrformurn(irregstem)

For irregulars, all form information is in the stem entry, so we need a function to create form urns directory from this.

Analysis

Exported functions

CitableParserBuilder.parsetoken — Function

Delegate to specific functions based on type's citable trait value.

parsetoken(s, x; data)

It is an error to invoke the parsetoken using types that are not a parser.

parsetoken(, s, x; data)

Citable parsers must implement parsetoken.

parsetoken(, s, x; data)

Parse String s by looking it up in a given dictionary.

Parse a single token to an array of Analysis or nothing.

parsetoken(tkn, parser; data)

CitableParserBuilder.parsewordlist — Function

Parse a list of tokens with a CitableParser.

parsewordlist(vocablist, p; data, countinterval)

Returns a Dict mapping strings to a (possibly empty) vector of Analysis objects.

Override implementation in CitableParser.

parsewordlist(vocablist, parser; data, countinterval)

Returns a Dict mapping strings to a (possibly empty) vector of Analysis objects.

CitableParserBuilder.parsedocument — Function

Use a CitableParser to parse a CitableTextCorpus with each citable node containing containg a single token.

parsedocument(doc, p; data)

Should return a list of AnalyzedTokens.

Override implementation in CitableParserBuilder.

parsedocument(doc, p; data, countinterval)

Should return a list of AnalyzedTokens.

CitableParserBuilder.parsecorpus — Function

Use a CitableParser to parse a CitableTextCorpus with each citable node containing containg a single token.

parsecorpus(c, p; data, countinterval)

Should return a list of AnalyzedTokens.

Override implementation in CitableParserBuilder.

parsecorpus(c, p; data, countinterval)

Should return a list of AnalyzedTokens.

Functions used internally in parsing via a FST

Kanones.stemsforlex — Function

Find stems for a given lexeme.

stemsforlex(lex, stems)

Kanones.fstinflpath — Function

Find full path to fst-inflon your system.

fstinflpath()

Kanones.echopath — Function

Find full path to echoon your system.

echopath()

Kanones.functionforcategory — Function

Dictionary of functions to use with each analytical type to determine a FormUrn from the FST description of form.

functionforcategory()

Kanones.parsefst — Function

Parse a string of FST output for a single token to a list of Analysis objects.

parsefst(fststring)

Kanones.applyparser — Function

Apply a parser to a token using fst-infl. Suppress standard error since this generates show-stopping IO demands if routed to terminal.

applyparser(tkn, parser)

Uses the external fst-infl command to apply parser to a list of vocabulary, and returns the string output of fst-infl.

applyparser(vocablist, parser)

Generation

Exported:

Kanones.generate — Function

Generate a string for the form of a given Analysis using a given dataset.

generate(analysis, kd)

Generate a string for a given form and lexeme using a given dataset.

generate(form, lex, kd)

Internal:

Kanones.generateuninflected — Function

Generate vector of possible strings for correctly accented form of uninflected token identified by analysis.

generateuninflected(analysis, kd; ortho)

Because uninflected forms record a full, accented string in their stem record, all we have to do is find matching stem records, and extract their form members.

Kanones.generatenoun — Function

Generate vector of possible strings for correctly accented form of a noun form identified by lexeme.

generatenoun(form, lex, kd)

Generate vector of possible strings for correctly accented form of a noun form identified by analysis.

generatenoun(analysis, kd)

Kanones.FstBuilder

The functions of the FstBuilder submodule compose expressions in SFST-PL, the language of the Stuttgart FST toolkit.

Kanones.FstBuilder.buildparser — Function

Read data from src and fstdir and compile a SFST parser in target.

buildparser(src, fstdir, target)
buildparser(src, fstdir, target, label; force)

Returns the path to the compiled binary greek.a which can then be used as an argument to the parsetoken function.

Kanones.FstBuilder.buildlexicon — Function

Read all stem data and compose lexicon.fst.

buildlexicon(src, target)

Arguments

src is a Kanones.Dataset.

target is the full path to the desination file lexicon.fst.

Kanones.FstBuilder.buildinflection — Function

Read all rules data and compose inflection.fst.

buildinflection(src, target)

Arguments

src is a Kanones.Dataset.

target is the full path to the desination file inflection.fst.

Kanones.FstBuilder.buildfinalfst — Function

Compose top-level transducer greek.fst.

buildfinalfst(target)

Arguments

target is the full path to the desination file greek.fst.

Kanones.FstBuilder.buildmakefile — Function

Compose makefile.

buildmakefile(target)

Arguments

src is a Kanones.Dataset.

target is the full path to the desination file makefile.

Kanones.FstBuilder.compilefst — Function

Compile binary parser with make.

compilefst(target)

Arguments

target is a directory with a makefile.

Kanones.FstBuilder.buildacceptor — Function

Compose acceptor.fst and write to file target.

buildacceptor(target)

Kanones.FstBuilder.uninflsquasher — Function

Compose transducer for filtering uninflected forms.

uninflsquasher()

Kanones.FstBuilder.nounsquasher — Function

Compose transducer for filtering nouns.

nounsquasher()

Kanones.FstBuilder.installalphabet — Function

Install alphabet.fst required to be part of an individual project.

installalphabet(src, target)

Kanones.FstBuilder.installsymbols — Function

Install symbols FST from generic FST source applying to many projects.

installsymbols(src, target)

Kanones.FstBuilder.symbolsfst — Function

Compose content for top-level symbols.fst file.

symbolsfst(dir)

Kanones.FstBuilder.fstok — Function

True if all necessary SFST components are installed.

fstok()

Kanones.FstBuilder.cmdok — Function

True if executable cmd is found.

cmdok(cmd)

Kanones.FstBuilder.fst — Function

Compose FST representation of a single UniflectedStem.

fst(stem, ortho)

Compose FST representation of a single UninflectedRule.

fst(rule, ortho)

Compose FST representation of a single NounStem.

fst(stem, ortho)

Compose FST representation of a single NounRule.

fst(rule, ortho)

Compose FST representation of a single PronounStem.

fst(stem, ortho)

Compose FST representation of a single PronounRule.

fst(rule, ortho)

Compose FST representation of a single AdjectiveStem.

fst(stem, ortho)

Compose FST representation of a single AdjectiveRule.

fst(rule, ortho)

Compose FST representation of a single IrregularNounStem.

fst(stem, ortho)

Compose FST representation of a single IrregularVerbStem.

fst(stem, ortho)

Compose FST representation of a single IrregularVerbStem.

fst(stem, ortho)

Compose FST representation of a single UninflectedRule.

fst(rule, ortho)

Compose FST representation of a single VerbStem.

fst(stem, ortho)

Compose FST representation of a single VerbRule.

fst(rule, ortho)

Compose FST representation of a single InfinitiveRule.

fst(rule, ortho)

Compose FST representation of a single ParticipleRule.

fst(rule, ortho)

Compose FST representation of a single InfinitiveRule.

fst(rule, ortho)

Kanones.FstBuilder.addsmooth — Function

Add smooth breathing to a vowel.

addsmooth(vowel)

Kanones.FstBuilder.addrough — Function

Add rough breathing to a vowel.

addrough(vowel)

Kanones.FstBuilder.greekfromfst — Function

Convert FST string to Unicode representation of Greek.

greekfromfst(fst)

Kanones.FstBuilder.fstgreek — Function

Convert Greek string s to FST representation.

fstgreek(s)
fstgreek(s, ortho)